2021-06-13 Roman (stable) - !!:FIX:SYS,BD.LDAP,Archive.{DBArch,FSArch},DAQ.{ModBus,MMS},{DAQ,Protocol}.OPC_UA,UI.Vision: Some cleaned from the build warnings. - SYS: The global variable limCacheIts_N type changed to "unsigned". - !FIX:{DAQ,Protocol}.OPC_UA.libOPC_UA: Fixed in the NodeId::operator=() function for comparing other string types,         like to Guid and Opaque.     FIX: Fixed in building on the Nokia N800 environment and the field "timestampsToReturn" type wrong conversion to REAL. - FIX:UI.WebVision: Fixed in form the configured width of the non string arguments of the primitive Text.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2753 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/arhiv/DBArch/mess.cpp
@@ -113,7 +113,7 @@ void ModMArch::start( )
 	reqEl.fldAdd(new TFld("TM",_("Time, seconds"),TFld::Integer,TCfg::Key|(tmAsStr()?TFld::DateTimeDec:0),(tmAsStr()?"20":"10")));
 	reqEl.fldAdd(new TFld("TMU",_("Time, microseconds"),TFld::Integer,TCfg::Key,"6","0"));
 	reqEl.fldAdd(new TFld("CATEG",_("Category"),TFld::String,TCfg::Key,"200"));
-	reqEl.fldAdd(new TFld("MESS",_("Message"),TFld::String,(keyTmCat()?TFld::NoFlag:TCfg::Key),(keyTmCat()?"100000":"255")));
+	reqEl.fldAdd(new TFld("MESS",_("Message"),TFld::String,(keyTmCat()?(int)TFld::NoFlag:(int)TCfg::Key),(keyTmCat()?"100000":"255")));
 	reqEl.fldAdd(new TFld("LEV",_("Level"),TFld::Integer,TFld::NoFlag,"2"));
     }
